Refactor pull up generic abstract class with context menu functionality.
